Migrate code base to java 17 (#463)
* Code cleanups for java 17

* Code cleanups for java 17

* Code cleanups for java 17

* Code cleanups for java 17

* Code cleanups for java 17

* Code cleanups for java 17

* Code cleanups for java 17

* Code cleanups for java 17

* Code cleanups for java 17

+++ app/src/main/java/com/github/frimtec/android/pikettassist/ui/alerts/AlertListFragment.java
@@ -51,7 +51,7 @@ import java.util.Optional;
 
 public class AlertListFragment extends AbstractListFragment<Alert> {
 
-  private static final String TAG = "AlertActivity";
+  private static final String TAG = "AlertListFragment";
 
   private static final int MENU_CONTEXT_VIEW_ID = 1;
   private static final int MENU_CONTEXT_DELETE_ID = 2;
@@ -156,7 +156,7 @@ public class AlertListFragment extends AbstractListFragment<Alert> {
 
   private long countAlertsWithinLastDays(List<Alert> alertList, int days) {
     Instant lastDaysStart = Instant.now().minus(days, DAYS);
-    return alertList.stream().filter(alert -> alert.getStartTime().isAfter(lastDaysStart)).count();
+    return alertList.stream().filter(alert -> alert.startTime().isAfter(lastDaysStart)).count();
   }
 
   @Override
@@ -173,22 +173,31 @@ public class AlertListFragment extends AbstractListFragment<Alert> {
   @Override
   public boolean onFragmentContextItemSelected(MenuItem item) {
     AdapterView.AdapterContextMenuInfo info = (AdapterView.AdapterContextMenuInfo) item.getMenuInfo();
+    if (info == null) {
+      Log.w(TAG, "No menu item was selected");
+      return false;
+    }
     ListView listView = getListView();
     Alert selectedAlert = (Alert) listView.getItemAtPosition(info.position);
     switch (item.getItemId()) {
-      case MENU_CONTEXT_VIEW_ID:
+      case MENU_CONTEXT_VIEW_ID -> {
         showAlertDetails(selectedAlert);
         return true;
-      case MENU_CONTEXT_DELETE_ID:
-        DialogHelper.areYouSure(getContext(), (dialog, which) -> {
-          this.alertDao.delete(selectedAlert);
-          refresh();
-          Toast.makeText(getContext(), R.string.general_entry_deleted, Toast.LENGTH_SHORT).show();
-        }, (dialog, which) -> {
-        });
+      }
+      case MENU_CONTEXT_DELETE_ID -> {
+        DialogHelper.areYouSure(
+            getContext(), (dialog, which) -> {
+              this.alertDao.delete(selectedAlert);
+              refresh();
+              Toast.makeText(getContext(), R.string.general_entry_deleted, Toast.LENGTH_SHORT).show();
+            }, (dialog, which) -> {
+            }
+        );
         return true;
-      default:
+      }
+      default -> {
         return false;
+      }
     }
   }
 
@@ -217,7 +226,7 @@ public class AlertListFragment extends AbstractListFragment<Alert> {
   private void showAlertDetails(Alert selectedAlert) {
     Intent intent = new Intent(this.getContext(), AlertDetailActivity.class);
     Bundle bundle = new Bundle();
-    bundle.putLong(AlertDetailActivity.EXTRA_ALERT_ID, selectedAlert.getId());
+    bundle.putLong(AlertDetailActivity.EXTRA_ALERT_ID, selectedAlert.id());
     intent.putExtras(bundle);
     startActivity(intent);
   }
